Fixed wrong result when dividing neg by pos
Before calling idiv, %edx was always being set to 0. That works fine for
positive numbers, but not for negative ones.

Example (the operand is stored in %eax):
	Passing 10 as the first operand of /
	%edx: 00...00
	%eax: 00...1010
	That works fine, %edx:%eax ends up being 10.

	Passing -10 as the first operand of /
	%edx: 00...00
	%eax: 11...10101
	That's clearly wrong. %edx:%eax is a very large positive number,
	nowhere near -10.

To fix this, we need to set %edx to all zeros if %eax is positive or all
ones if it's negative. The instruction cltd does exactly that
(apparently cltd stands for "convert long to double long"), it
sign-extends %eax into %edx.